Refactored several methods for ner MerchantCustomizations functionality. A few methods still remain.

+++ src/main/java/model/Armor.java
@@ -1,7 +1,7 @@
 package model;
 
 public class Armor extends Weapon {
-	public Armor(String name, double marketValue, String rarity, String type) {
+	public Armor(String name, int marketValue, String rarity, String type) {
 		super(name, marketValue, rarity, type);
 	}
 }
